+++ src/ScreenCover/index.js
@@ -1,26 +1,62 @@
-import React, { Fragment } from 'react'
+import React from 'react'
 import propTypes from 'prop-types'
 import css from 'styled-jsx/css'
 import cx from 'classnames'
 
-export const ScreenCover = ({ onClick, withoutBackgroundColor }) => (
-    <div
-        className={cx({
-            'with-background-color': !withoutBackgroundColor,
-        })}
-        onClick={onClick}
-    >
+const Backdrop = ({ onClick, withoutBackgroundColor }) => (
+    <div className="backdrop" onClick={onClick}>
         <style jsx>{`
             div {
                 height: 100%;
-                left: 0;
-                position: absolute;
-                top: 0;
                 width: 100%;
-            }
 
-            .with-background-color {
-                background: rgba(200, 200, 200, 0.6);
+                background: ${withoutBackgroundColor
+                    ? 'none'
+                    : 'rgba(200, 200, 200, 0.6)'};
+            }
+        `}</style>
+    </div>
+)
+
+const Content = ({ children }) => (
+    <div>
+        {children}
+        <style jsx>{`
+            div {
+                position: absolute;
+                top: 50%;
+                left: 50%;
+                width: auto;
+                height: auto;
+                transform: translate(-50%, -50%);
+            }
+        `}</style>
+    </div>
+)
+
+const ScreenCover = ({
+    children,
+    onClick,
+    className,
+    withoutBackgroundColor,
+}) => (
+    <div className={className}>
+        <Backdrop
+            onClick={onClick}
+            withoutBackgroundColor={withoutBackgroundColor}
+        />
+        <Content>{children}</Content>
+
+        <style jsx>{`
+            div {
+                position: fixed;
+                height: 100%;
+                width: 100%;
+
+                left: 0;
+                top: 0;
+
+                z-index: 10000;
             }
         `}</style>
     </div>
@@ -28,5 +64,9 @@ export const ScreenCover = ({ onClick, withoutBackgroundColor }) => (
 
 ScreenCover.propTypes = {
     onClick: propTypes.func,
+    className: propTypes.string,
+    children: propTypes.node,
     withoutBackgroundColor: propTypes.bool,
 }
+
+export { ScreenCover }