+++ stories/CircularProgress.stories.js
@@ -5,14 +5,6 @@ import { CircularProgress } from '../src'
 storiesOf('CircularProgress', module)
     .add('Default', () => <CircularProgress />)
 
-    .add('Overlay component', () => (
-        <div style={{ width: '358px', height: '358px', position: 'relative' }}>
-            <CircularProgress overlay />
-        </div>
-    ))
-
-    .add('Overlay page', () => <CircularProgress overlay />)
-
     .add('Small', () => <CircularProgress small />)
 
     .add('Large', () => <CircularProgress large />)